FSA-4709: Minor code improvements, added if check

+++ projects/fsastorefrontlib/src/cms-components/my-account/claim/delete-claim-dialog/delete-claim-dialog.component.ts
@@ -39,9 +39,11 @@ export class DeleteClaimDialogComponent implements OnInit {
     this.authService
       .getOccUserId()
       .pipe(take(1))
-      .subscribe(occUserId =>
-        this.service.removeClaim(occUserId, this.claimNumber)
-      )
+      .subscribe(occUserId => {
+        if (occUserId) {
+          this.service.removeClaim(occUserId, this.claimNumber);
+        }
+      })
       .unsubscribe();
   }
 }
